rename: aside tab icon 컴포넌트 이름 수정
