+++ src/components/aside/aside.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useNavigate, useLocation } from 'react-router';
 
 import UserProfile from './user-profile';
-import { Asidetab } from './aside-tab-icon';
+import { Asidetab } from './aside-tab-menu';
 
 import { MessageCircleMore } from 'lucide-react';
